Android: Migrated to new Gradle build sysyem
