+++ .gitignore
@@ -23,3 +23,4 @@ dep/libsoldout-1.3/mkd2latex
 dep/libsoldout-1.3/mkd2man
 test/*.tst
 *.class
+dep/libsoldout/depends
