+++ platform/android/build.gradle
@@ -1,0 +1,12 @@
+allprojects {
+
+    repositories {
+        jcenter()
+    }
+
+    buildscript {
+        repositories {
+            jcenter()
+        }
+    }
+}
