--- platform/android/functional-test-tests/res/values/strings.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <string name="app_name">Bypass-Test-FrameworkTest</string>
-
-</resources>
