--- platform/android/library/AndroidManifest.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-      package="in.uncod.android.bypass"
-      android:versionCode="1"
-      android:versionName="1.0">
-
-    <uses-sdk
-        android:minSdkVersion="8"
-        android:targetSdkVersion="16" />
-
-    <application android:label="@string/app_name" >
-    </application>
-</manifest>
