--- platform/android/library/jni/Application.mk
@@ -1,2 +1,0 @@
-APP_ABI := armeabi armeabi-v7a arm64-v8a x86 x86_64 mips mips64
-APP_STL := gnustl_static
