--- platform/android/library/pom.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId>in.uncod.android.bypass</groupId>
-	<artifactId>bypass</artifactId>
-	<packaging>apklib</packaging>
-	<name>Bypass</name>
-
-	<parent>
-		<groupId>in.uncod.android.bypass</groupId>
-		<artifactId>parent</artifactId>
-		<version>1.2-SNAPSHOT</version>
-		<relativePath>../pom.xml</relativePath>
-	</parent>
-
-	<dependencies>
-		<dependency>
-			<groupId>com.google.android</groupId>
-			<artifactId>android</artifactId>
-			<scope>provided</scope>
-		</dependency>
-	</dependencies>
-
-	<build>
-		<sourceDirectory>src</sourceDirectory>
-
-		<plugins>
-			<plugin>
-				<groupId>com.jayway.maven.plugins.android.generation2</groupId>
-				<artifactId>android-maven-plugin</artifactId>
-				<executions>
-					<!-- Add the execution of the ndk-build goal - this will compile the 
-						native code in addition to building the Java & APK -->
-					<execution>
-						<goals>
-							<goal>ndk-build</goal>
-						</goals>
-						<configuration>
-							<target>all</target>
-							<deleteConflictingFiles>true</deleteConflictingFiles>
-							<!-- Configure the plugin to remove all built artifacts from the libs/armeabi 
-								folder when the native build is complete -->
-							<clearNativeArtifacts>true</clearNativeArtifacts>
-							<!-- Configure the plugin to also attach the built, native, artifact 
-								from the libs/armeabi folder when the native build is complete. This puts 
-								the artifact into the Maven repository -->
-							<attachNativeArtifacts>true</attachNativeArtifacts>
-						</configuration>
-					</execution>
-				</executions>
-				<extensions>true</extensions>
-			</plugin>
-		</plugins>
-	</build>
-</project>
