+++ platform/android/library/src/main/AndroidManifest.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="in.uncod.android.bypass"
+    android:versionCode="1"
+    android:versionName="1.0">
+
+    <application></application>
+</manifest>
