+++ platform/android/library/src/main/java/in/uncod/android/bypass/Bypass.java
