+++ platform/android/library/src/main/jni/Application.mk
@@ -1,0 +1,4 @@
+APP_ABI := armeabi armeabi-v7a arm64-v8a x86 x86_64 mips mips64
+APP_STL := gnustl_static
+
+NDK_APP_OUT := ../../../build/ndk-obj
