+++ platform/android/library/src/main/jni/bypass.cpp
