+++ platform/android/sample/src/main/res/drawable-hdpi/ic_launcher.png
