+++ platform/android/sample/src/main/res/drawable-mdpi/ic_launcher.png
