+++ platform/android/sample/src/main/res/drawable-xhdpi/ic_launcher.png
