+++ platform/android/sample/src/main/res/layout/main.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:orientation="vertical"> 
+
+    <TextView android:id="@+id/demoText"
+        android:layout_width="fill_parent"
+	android:layout_height="wrap_content"
+	android:layout_margin="5dp"/>
+    
+</ScrollView>
