+++ platform/android/sample/src/main/res/raw/readme.txt
@@ -1,0 +1,37 @@
+#Bypass
+
+#Header sizes
+##Smaller header
+##Even smaller header
+
+Paragraphs are obviously supported along with all the fancy text styling you could want.
+There is *italic*, **bold** and ***bold italic***. Even links are supported, visit the
+github page for Bypass [here](https://github.com/Uncodin/bypass).
+
+* Nested List
+	* One
+	* Two
+	* Three
+* One
+	* One
+	* Two
+	* Three
+
+## Code Block Support
+
+	const char* str;
+	str = env->GetStringUTFChars(markdown, NULL);
+	
+	Bypass::Parser parser;
+	Bypass::Document document = parser.parse(str);
+	
+	env->ReleaseStringUTFChars(markdown, str);
+	
+
+#Quote Support
+> Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam ullamcorper tempus neque vulputate congue.
+> Quisque ipsum tellus, pharetra vitae dictum sit amet, placerat vitae felis. Etiam ornare eleifend libero,
+> et dignissim nibh rhoncus eu. Morbi mattis nisi in velit tristique eu mollis odio molestie.
+> Phasellus pretium dapibus mattis. Nullam mollis nisl id enim vestibulum quis lacinia turpis pellentesque. 
+> Integer id mi quam. Fusce sagittis cursus eros, nec convallis ligula consectetur quis. Suspendisse potenti.
+
