+++ platform/android/sample/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<resources>
+
+    <string name="app_name">Bypass-Test</string>
+
+</resources>
