+++ platform/android/settings.gradle
@@ -1,0 +1,2 @@
+include ':library'
+include ':sample'
